[PoC][API] Separate admin and shop docs
