+++ src/Sylius/Bundle/ApiBundle/Resources/config/app/config.yaml
@@ -5,6 +5,8 @@ api_platform:
     patch_formats:
         json: ['application/merge-patch+json']
     path_segment_name_generator: api_platform.path_segment_name_generator.dash
+    enable_swagger_ui: false
+    enable_re_doc: false
     swagger:
         versions: [3]
         api_keys:
