+++ src/Sylius/Bundle/ApiBundle/Swagger/DocumentationNormalizer.php
@@ -1,0 +1,179 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Bundle\ApiBundle\Swagger;
+
+use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+
+final class DocumentationNormalizer implements NormalizerInterface
+{
+    /** @var NormalizerInterface */
+    private $decoratedNormalizer;
+
+    /** @var RequestStack */
+    private $requestStack;
+
+    public function __construct(NormalizerInterface $decoratedNormalizer, RequestStack $requestStack)
+    {
+        $this->decoratedNormalizer = $decoratedNormalizer;
+        $this->requestStack = $requestStack;
+    }
+
+    public function supportsNormalization($data, string $format = null): bool
+    {
+        return $this->decoratedNormalizer->supportsNormalization($data, $format);
+    }
+
+    public function normalize($object, string $format = null, array $context = [])
+    {
+        $docs = $this->decoratedNormalizer->normalize($object, $format, $context);
+
+        if ($this->isAdminDocs()) {
+            $docs['paths'] = array_filter($docs['paths']->getArrayCopy(), function ($path) {
+                return substr($path, 0, 14) === '/new-api/admin';
+            }, ARRAY_FILTER_USE_KEY);
+
+            $docs['components']['schemas']['AdminUserToken'] = [
+                'type' => 'object',
+                'properties' => [
+                    'token' => [
+                        'type' => 'string',
+                        'readOnly' => true,
+                    ],
+                ],
+            ];
+
+            $docs['components']['schemas']['AdminUserCredentials'] = [
+                'type' => 'object',
+                'properties' => [
+                    'email' => [
+                        'type' => 'string',
+                        'example' => 'api@example.com',
+                    ],
+                    'password' => [
+                        'type' => 'string',
+                        'example' => 'sylius-api',
+                    ],
+                ],
+            ];
+
+            $tokenDocumentation = [
+                'paths' => [
+                    '/new-api/admin/authentication-token' => [
+                        'post' => [
+                            'tags' => ['AdminUserToken'],
+                            'operationId' => 'postCredentialsItem',
+                            'summary' => 'Get JWT token to login.',
+                            'requestBody' => [
+                                'description' => 'Create new JWT Token',
+                                'content' => [
+                                    'application/json' => [
+                                        'schema' => [
+                                            '$ref' => '#/components/schemas/AdminUserCredentials',
+                                        ],
+                                    ],
+                                ],
+                            ],
+                            'responses' => [
+                                Response::HTTP_OK => [
+                                    'description' => 'Get JWT token',
+                                    'content' => [
+                                        'application/json' => [
+                                            'schema' => [
+                                                '$ref' => '#/components/schemas/AdminUserToken',
+                                            ],
+                                        ],
+                                    ],
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ];
+
+            return array_merge_recursive($tokenDocumentation, $docs);
+        }
+
+        $docs['paths'] = array_filter($docs['paths']->getArrayCopy(), function ($path) {
+            return substr($path, 0, 13) === '/new-api/shop';
+        }, ARRAY_FILTER_USE_KEY);
+
+        $docs['components']['schemas']['ShopUserToken'] = [
+            'type' => 'object',
+            'properties' => [
+                'token' => [
+                    'type' => 'string',
+                    'readOnly' => true,
+                ],
+            ],
+        ];
+
+        $docs['components']['schemas']['ShopUserCredentials'] = [
+            'type' => 'object',
+            'properties' => [
+                'email' => [
+                    'type' => 'string',
+                    'example' => 'shop@example.com',
+                ],
+                'password' => [
+                    'type' => 'string',
+                    'example' => 'sylius',
+                ],
+            ],
+        ];
+
+        $tokenDocumentation = [
+            'paths' => [
+                '/new-api/shop/authentication-token' => [
+                    'post' => [
+                        'tags' => ['ShopUserToken'],
+                        'operationId' => 'postCredentialsItem',
+                        'summary' => 'Get JWT token to login.',
+                        'requestBody' => [
+                            'description' => 'Create new JWT Token',
+                            'content' => [
+                                'application/json' => [
+                                    'schema' => [
+                                        '$ref' => '#/components/schemas/ShopUserCredentials',
+                                    ],
+                                ],
+                            ],
+                        ],
+                        'responses' => [
+                            Response::HTTP_OK => [
+                                'description' => 'Get JWT token',
+                                'content' => [
+                                    'application/json' => [
+                                        'schema' => [
+                                            '$ref' => '#/components/schemas/ShopUserToken',
+                                        ],
+                                    ],
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ];
+
+        return array_merge_recursive($tokenDocumentation, $docs);
+    }
+
+    private function isAdminDocs(): bool
+    {
+        return $this->requestStack->getMasterRequest()->getRequestUri() === '/new-api/admin/docs';
+
+    }
+}
