--- src/Sylius/Bundle/ApiBundle/Swagger/ShopAuthenticationTokenDocumentationNormalizer.php
@@ -1,98 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Sylius\Bundle\ApiBundle\Swagger;
-
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
-
-final class ShopAuthenticationTokenDocumentationNormalizer implements NormalizerInterface
-{
-    /** @var NormalizerInterface */
-    private $decoratedNormalizer;
-
-    public function __construct(NormalizerInterface $decoratedNormalizer)
-    {
-        $this->decoratedNormalizer = $decoratedNormalizer;
-    }
-
-    public function supportsNormalization($data, string $format = null): bool
-    {
-        return $this->decoratedNormalizer->supportsNormalization($data, $format);
-    }
-
-    public function normalize($object, string $format = null, array $context = [])
-    {
-        $docs = $this->decoratedNormalizer->normalize($object, $format, $context);
-
-        $docs['components']['schemas']['ShopUserToken'] = [
-            'type' => 'object',
-            'properties' => [
-                'token' => [
-                    'type' => 'string',
-                    'readOnly' => true,
-                ],
-            ],
-        ];
-
-        $docs['components']['schemas']['ShopUserCredentials'] = [
-            'type' => 'object',
-            'properties' => [
-                'email' => [
-                    'type' => 'string',
-                    'example' => 'shop@example.com',
-                ],
-                'password' => [
-                    'type' => 'string',
-                    'example' => 'sylius',
-                ],
-            ],
-        ];
-
-        $tokenDocumentation = [
-            'paths' => [
-                '/new-api/shop/authentication-token' => [
-                    'post' => [
-                        'tags' => ['ShopUserToken'],
-                        'operationId' => 'postCredentialsItem',
-                        'summary' => 'Get JWT token to login.',
-                        'requestBody' => [
-                            'description' => 'Create new JWT Token',
-                            'content' => [
-                                'application/json' => [
-                                    'schema' => [
-                                        '$ref' => '#/components/schemas/ShopUserCredentials',
-                                    ],
-                                ],
-                            ],
-                        ],
-                        'responses' => [
-                            Response::HTTP_OK => [
-                                'description' => 'Get JWT token',
-                                'content' => [
-                                    'application/json' => [
-                                        'schema' => [
-                                            '$ref' => '#/components/schemas/ShopUserToken',
-                                        ],
-                                    ],
-                                ],
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ];
-
-        return array_merge_recursive($docs, $tokenDocumentation);
-    }
-}
